requirements.txt
# AI Operations Assistant Dependencies

# LLM - Google Gemini
google-generativeai>=0.3.0

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# HTTP Requests
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Development (optional)
pytest>=7.4.0
black>=24.1.0
flake8>=7.0.0
